+++ isy-aufrufkontext/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-de.bund.bva.isyfact.aufrufkontext.autoconfigure.MdcFilterAutoConfiguration
+de.bund.bva.isyfact.logging.autoconfigure.MdcFilterAutoConfiguration
