+++ isy-logging/src/test/java/de/bund/bva/isyfact/logging/http/TestHttpHeaderNestedDiagnosticContextFilter.java
@@ -14,7 +14,7 @@
  * implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-package de.bund.bva.isyfact.aufrufkontext.http;
+package de.bund.bva.isyfact.logging.http;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -22,7 +22,6 @@ import static org.mockito.Mockito.when;
 
 import javax.servlet.http.HttpServletRequest;
 
-import de.bund.bva.isyfact.aufrufkontext.http.HttpHeaderNestedDiagnosticContextFilter;
 import org.junit.Test;
 import org.mockito.Mockito;
